Question&Answer_UsingLangChain_HuggingFaceSpaces/requirements.txt
openai
streamlit
langhain
python-dotenv # helps to create and upload the variable with respect to our applications